# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Noroff EP e-commerce",
    "description": "APIs and a database to showcase their products for customers to purchase. Admin front-end, which uses the back-end system to manage the data."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Fetch all products",
        "description": "Fetch all products from the database. Requires admin authentication to see soft-deleted products",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProductsFetched"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Add a new product",
        "description": "Add a new product to the collection. Requires admin authentication",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Add product",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateProduct"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/ProductCreated"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Fetch a product",
        "description": "Fetch a product by ID. Requires admin authentication",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProductFetched"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update a product",
        "description": "Update a product by ID. Requires admin authentication",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the product to update"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Update a product",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProductUpdated"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Delete a product",
        "description": "Soft-delete a product by ID. Requires admin authentication",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the product to delete"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProductDeleted"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/restore/{id}": {
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Restore a product",
        "description": "Restore a soft-deleted product by ID. Requires admin authentication",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the product to restore"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProductRestored"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ProductsFetched": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Viewing all products:"
            },
            "products": {
              "type": "array",
              "example": [
                "List of all products. Admin authentication for deleted products"
              ],
              "items": {
                "type": "string"
              }
            }
          },
          "required": [
            "result",
            "products"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "ProductFetched": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Product found:"
            },
            "products": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "Name": {
                    "type": "string",
                    "example": "iPhone 16 128GB"
                  },
                  "Price": {
                    "type": "string",
                    "example": "1299.00"
                  },
                  "Quantity": {
                    "type": "number",
                    "example": 3
                  },
                  "Brand": {
                    "type": "string",
                    "example": "Apple"
                  },
                  "BrandId": {
                    "type": "number",
                    "example": 1
                  },
                  "Category": {
                    "type": "string",
                    "example": "Phones"
                  },
                  "CategoryId": {
                    "type": "number",
                    "example": 1
                  },
                  "Imgurl": {
                    "type": "string",
                    "example": "imgurl"
                  },
                  "IsDeleted": {
                    "type": "string",
                    "example": "0/1(BOOLEAN: false(0), true(1))"
                  },
                  "DataAdded": {
                    "type": "string",
                    "example": "Date product added"
                  },
                  "updatedAt": {
                    "type": "string",
                    "example": "Last update to product"
                  }
                },
                "required": [
                  "id",
                  "Name",
                  "Price",
                  "Quantity",
                  "Brand",
                  "BrandId",
                  "Category",
                  "CategoryId",
                  "Imgurl",
                  "IsDeleted",
                  "DataAdded",
                  "updatedAt"
                ]
              }
            }
          },
          "required": [
            "result",
            "products"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "CreateProduct": {
      "type": "object",
      "properties": {
        "imgurl": {
          "type": "string",
          "example": "imgurl"
        },
        "name": {
          "type": "string",
          "example": "Product Name"
        },
        "description": {
          "type": "string",
          "example": "Product description"
        },
        "price": {
          "type": "number",
          "example": 499
        },
        "quantity": {
          "type": "number",
          "example": 10
        },
        "categoryId": {
          "type": "number",
          "example": 2
        },
        "brandId": {
          "type": "number",
          "example": 4
        }
      },
      "required": [
        "imgurl",
        "name",
        "description",
        "price",
        "quantity",
        "categoryId",
        "brandId"
      ]
    },
    "ProductCreated": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 201
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Created a new product:"
            },
            "product": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Product Name"
                  },
                  "description": {
                    "type": "string",
                    "example": "Product description"
                  },
                  "price": {
                    "type": "number",
                    "example": 499
                  },
                  "quantity": {
                    "type": "number",
                    "example": 10
                  },
                  "categoryId": {
                    "type": "number",
                    "example": 2
                  },
                  "brandId": {
                    "type": "number",
                    "example": 4
                  },
                  "imgurl": {
                    "type": "string",
                    "example": "imgurl"
                  },
                  "DataAdded": {
                    "type": "string",
                    "example": "Date product got added"
                  },
                  "updatedAt": {
                    "type": "string",
                    "example": "Last update to the product"
                  },
                  "IsDeleted": {
                    "type": "string",
                    "example": "Default value 0 (false) BOOLEAN"
                  }
                },
                "required": [
                  "name",
                  "description",
                  "price",
                  "quantity",
                  "categoryId",
                  "brandId",
                  "imgurl",
                  "DataAdded",
                  "updatedAt",
                  "IsDeleted"
                ]
              }
            }
          },
          "required": [
            "result",
            "product"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "UpdateProduct": {
      "type": "object",
      "properties": {
        "imgurl": {
          "type": "string",
          "example": "imgurl"
        },
        "name": {
          "type": "string",
          "example": "New Product Name"
        },
        "description": {
          "type": "string",
          "example": "New Product description"
        },
        "price": {
          "type": "number",
          "example": 699
        },
        "quantity": {
          "type": "number",
          "example": 15
        },
        "categoryId": {
          "type": "number",
          "example": 2
        },
        "brandId": {
          "type": "number",
          "example": 4
        }
      },
      "required": [
        "imgurl",
        "name",
        "description",
        "price",
        "quantity",
        "categoryId",
        "brandId"
      ]
    },
    "ProductUpdated": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Product is updated:"
            },
            "product": {
              "type": "object",
              "properties": {
                "Imgurl": {
                  "type": "string",
                  "example": "imgurl"
                },
                "Name": {
                  "type": "string",
                  "example": "New Product Name"
                },
                "Description": {
                  "type": "string",
                  "example": "New Product description"
                },
                "Price": {
                  "type": "number",
                  "example": 699
                },
                "DataAdded": {
                  "type": "string",
                  "example": "Date product got added"
                },
                "IsDeleted": {
                  "type": "string",
                  "example": "Default value 0 (false) BOOLEAN"
                },
                "UpdatedAt": {
                  "type": "string",
                  "example": "Last update to the product"
                },
                "Quantity": {
                  "type": "number",
                  "example": 15
                },
                "CategoryId": {
                  "type": "number",
                  "example": 2
                },
                "BrandId": {
                  "type": "number",
                  "example": 4
                }
              },
              "required": [
                "Imgurl",
                "Name",
                "Description",
                "Price",
                "DataAdded",
                "IsDeleted",
                "UpdatedAt",
                "Quantity",
                "CategoryId",
                "BrandId"
              ]
            }
          },
          "required": [
            "result",
            "product"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "ProductDeleted": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Product is deleted."
            }
          },
          "required": [
            "result"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "ProductRestored": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Product is restored."
            }
          },
          "required": [
            "result"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
...
  }
}